Extract tester_invocation.rs from turn.rs (parent #680)
Fourth vertical-slice extraction: move the post-loop tester smoke-test
invocation flow out of turn.rs into a new
src/agent/loop_run/tester_invocation.rs sibling module.

Migrated (all free fns over &mut Agent / &Agent):

Entry point (pub(super)):
- try_invoke_tester(&mut Agent, &[String]) -> bool — invoked from
  success.rs after VerifierOutcome::TesterDelegated.

Private helpers:
- log_tester_event / log_tester_skip (telemetry)
- detect_tester_candidate (candidate filtering)
- ensure_tester_runs_root (mkdir + telemetry on failure)
- handle_tester_outcome (record_feedback_if_unset + telemetry)

No facade re-export (DR3-001). Caller updates:
- success.rs: `self.try_invoke_tester(&stats.changed_files)` →
  `super::tester_invocation::try_invoke_tester(self, &stats.changed_files)`

`cargo fix --tests` pruned 1 newly-unused import from turn.rs's prelude.

turn.rs: 6,123 -> 5,945 (-178 LOC)
Cumulative from 39,489: -33,544 LOC (-84.9%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
