+++ src/agent/loop_run/turn.rs
@@ -54,7 +54,6 @@ use super::small_helpers::{
 };
 use super::spinner::{Spinner, SpinnerStopSignal};
 use super::summary::{ExitReason, LoopResult};
-use super::tester;
 use super::tool_display::progress_path_display;
 use super::tool_execution::{
     failed_outcome_for_call, rejected_outcome_for_call, success_outcome_for_call,
@@ -1397,184 +1396,6 @@ impl Agent {
             );
         }
     }
-
-    /// Issue #459: try to invoke the Tester Skill when `AutoTestRunner::detect`
-    /// returned None. Returns `true` iff the Tester recorded a FeedbackFrame
-    /// (so the caller skips the `NoVerifierAvailable` fallback). Disable
-    /// gating (Plan / `ANVIL_NO_TESTER` / per-turn cap / no candidate) is
-    /// evaluated here so the orchestrator (`run_tester_with_strategy`) only
-    /// sees the run-body inputs.
-    ///
-    /// `Aborted` outcomes (LLM malformed / approval denied / harness build
-    /// failure) consume the per-turn cap and **do not** record a frame —
-    /// caller falls through to the `NoVerifierAvailable` fallback per design
-    /// § 4-2 ("Skip / Abort の細粒度 variant は同じ branch (= 既存
-    /// no_verifier) に集約し、log のみで識別する"). This is the boundary
-    /// captured by the bool return.
-    pub(super) fn try_invoke_tester(&mut self, changed_files: &[String]) -> bool {
-        if let Some(reason) = tester::check_invocation_gate(
-            self.tester_called_this_turn,
-            self.session.mode_state.mode == ExecutionMode::Plan,
-            tester::tester_disabled(|key| std::env::var(key).ok()),
-        ) {
-            self.log_tester_skip(reason.as_str());
-            return false;
-        }
-        let candidate = match self.detect_tester_candidate(changed_files) {
-            Some(candidate) => candidate,
-            None => return false,
-        };
-        let session_dir = self
-            .session_store
-            .state_root()
-            .join("sessions")
-            .join(self.session_store.session_id());
-        let tmp_tests_root = session_dir.join("tmp-tests");
-        let tester_runs_root = session_dir.join("tester-runs");
-        if !self.ensure_tester_runs_root(&tester_runs_root) {
-            return false;
-        }
-        let approval_mode =
-            tester::tester_approval_mode(self.config.yes_mode, io::stdin().is_terminal());
-        self.tester_called_this_turn = true;
-
-        let work_root = self.work_root.clone();
-        let session_id = self.session_store.session_id().to_string();
-        let run = tester::TesterRun {
-            work_root: &work_root,
-            tmp_tests_root: &tmp_tests_root,
-            tester_runs_root: &tester_runs_root,
-            approval_mode,
-            plan_mode: false,
-            no_tester_env: false,
-            session_id: std::borrow::Cow::Owned(session_id.clone()),
-        };
-
-        let session_id_for_log = session_id.clone();
-        let tester_client = self.client.clone();
-        let tester_main_model = self.models.main.clone();
-        let llm_call =
-            move |prompt: &tester::TesterPrompt| -> Result<String, tester::TesterLlmError> {
-                tester::run_tester_llm_call(
-                    &tester_client,
-                    &tester_main_model,
-                    &session_id_for_log,
-                    prompt,
-                )
-            };
-
-        let offline = self.config.offline;
-        let run_bash = move |cmd: &str,
-                             cwd: &Path,
-                             timeout: Option<std::time::Duration>|
-              -> Result<crate::tools::bash::BashExecutionOutcome, String> {
-            // No cancel_flag propagation: Tester's smoke run sits past the
-            // main interrupt monitor scope (post-loop hook). The 30s
-            // explicit_timeout still caps wall time.
-            //
-            // CB-003 (Issue #459): pass `BashEnvPolicy::TesterSanitized` so
-            // LLM-generated smoke code cannot read parent-process secrets
-            // (`OPENAI_API_KEY`, `GITHUB_TOKEN`, `AWS_*`, anything `*_TOKEN`/
-            // `*_SECRET`/`*_PASSWORD`). Only the explicit allowlist in
-            // `bash::TESTER_ENV_ALLOWLIST_EXACT` is forwarded.
-            crate::tools::bash::run_with_outcome(
-                cmd,
-                cwd,
-                None,
-                offline,
-                timeout,
-                Some(crate::tools::bash::BashEnvPolicy::TesterSanitized),
-            )
-            .map(|(_, outcome)| outcome)
-        };
-
-        let approver = move |mode: tester::ApprovalMode,
-                             command: &[String]|
-              -> Result<(), tester::AbortReason> {
-            // Honour the same write/run/promote 3-gate symmetry: Auto bypass /
-            // Forbidden deny / Interactive y/N. Production prompt goes through
-            // `prompt_for_approval(stdout, stdin)` so both ends are real TTY
-            // streams; CI takes the Forbidden branch above.
-            let mut stdout = std::io::stdout().lock();
-            let stdin_handle = std::io::stdin();
-            let mut stdin = stdin_handle.lock();
-            tester::prompt_for_approval(mode, command, &mut stdout, &mut stdin)
-        };
-
-        let outcome =
-            tester::run_tester_with_strategy(run, candidate, llm_call, run_bash, approver);
-        self.handle_tester_outcome(outcome, &session_id)
-    }
-
-    fn log_tester_event(&self, event: &'static str, payload: serde_json::Value) {
-        log_llm_event(event, payload);
-    }
-
-    fn log_tester_skip(&self, reason: &str) {
-        self.log_tester_event(
-            "agent.tester.skipped",
-            serde_json::json!({
-                "session_id": self.session_store.session_id(),
-                "skip_reason": reason,
-            }),
-        );
-    }
-
-    fn detect_tester_candidate(&self, changed_files: &[String]) -> Option<tester::TesterCandidate> {
-        let candidate = tester::TesterCandidate::detect(&self.work_root, changed_files);
-        if candidate.is_none() {
-            self.log_tester_skip(tester::NotInvokedReason::NoCandidate.as_str());
-        }
-        candidate
-    }
-
-    fn ensure_tester_runs_root(&self, tester_runs_root: &Path) -> bool {
-        if let Err(err) = std::fs::create_dir_all(tester_runs_root) {
-            self.log_tester_event(
-                "agent.tester.failed",
-                serde_json::json!({
-                    "session_id": self.session_store.session_id(),
-                    "failure_reason": format!("mkdir tester-runs: {err}"),
-                }),
-            );
-            return false;
-        }
-        true
-    }
-
-    fn handle_tester_outcome(&mut self, outcome: tester::TesterOutcome, session_id: &str) -> bool {
-        match outcome {
-            tester::TesterOutcome::Recorded(frame) => {
-                let kind_value =
-                    serde_json::to_value(&frame.kind).unwrap_or(serde_json::Value::Null);
-                self.session.record_feedback_if_unset(frame);
-                self.log_tester_event(
-                    "agent.tester.completed",
-                    serde_json::json!({
-                        "session_id": session_id,
-                        "feedback_kind": kind_value,
-                    }),
-                );
-                true
-            }
-            tester::TesterOutcome::NotInvoked(reason) => {
-                self.log_tester_skip(reason.as_str());
-                false
-            }
-            tester::TesterOutcome::Aborted(reason) => {
-                self.log_tester_event(
-                    "agent.tester.failed",
-                    serde_json::json!({
-                        "session_id": session_id,
-                        "failure_reason": reason.as_str(),
-                        "detail": reason.detail().map(|d| tester::sanitize_tester_log(d, tester::TESTER_LOG_CAP)),
-                    }),
-                );
-                false
-            }
-        }
-    }
-
     pub(super) fn request_assistant_reply_with_retry(
         &mut self,
         stream_output: bool,
